reducer cleaned, functions now imported into reducer
