+++ src/data/reducer/addPlayer.js
@@ -1,0 +1,6 @@
+export const addPlayer = (state, { data }) => {
+    return {
+        ...state,
+        players: [...state.players, { name: data.playerName, skill: data.playerSkill }]
+    };
+};
