+++ src/data/reducer/setNumberOfPlayers.js
@@ -1,0 +1,6 @@
+export const setNumberOfPlayers = (state, { value }) => {
+    return {
+        ...state,
+        numberOfPlayers: state.numberOfPlayers + value,
+    };
+};
