+++ src/data/reducer/toggleHomeTeam.js
@@ -1,0 +1,10 @@
+export const toggleHomeTeam = (state) => {
+    const homeTeam = state.team1Home;
+
+    const toggleTeam = homeTeam ? false : true;
+
+    return {
+        ...state,
+        team1Home: toggleTeam,
+    };
+};
