+++ src/data/reducer/toggleInstructions.js
@@ -1,0 +1,6 @@
+export const toggleInstructions = (state) => {
+    return {
+        ...state,
+        showInstructions: !state.showInstructions,
+    };
+};
